change attestation to coco
