+++ tools/xl/xl_cmdtable.c
@@ -641,10 +641,10 @@ const struct cmd_spec cmd_table[] = {
       "Issue a qemu monitor command to the device model of a domain",
       "<Domain> <Command>",
     },
-    { "attestation",
-      &main_attestation, 0, 0,
-      "request attestation for a coco and put it into a file",
-      "<options> <Domain>",
+    { "coco",
+      &main_coco, 0, 0,
+      "All subcommand for coco domains",
+      "<Options> <Domain> <Operation>",
     },
 #ifdef LIBXL_HAVE_DT_OVERLAY
     { "dt-overlay",
