+++ tools/xl/xl_misc.c
@@ -364,55 +364,6 @@ int main_config_update(int argc, char **argv)
     return 0;
 }
 
-int main_attestation(int argc, char **argv) {
-    int rc;
-    int dst_file = 1;
-    char * mmonce = NULL;
-    uint32_t domid;
-    bool is_mmonce_file = false;
-
-    int opt;
-    static struct option opts[] = {
-        {"file", 1, 0, 'f'},
-        {"print", 0, 0, 'p'},
-        {"mmonce", 1, 0, 'm'},
-        {"mmonce-file", 1, 0, 'n'},
-        COMMON_LONG_OPTS
-    };
-
-    SWITCH_FOREACH_OPT(opt, "f:pm:n:", opts, "attestation", 0) {
-    case 'p':
-        dst_file = 1;
-        break;
-    case 'f':
-        dst_file = open(optarg, O_WRONLY | O_CREAT, 0644);
-        if (!dst_file) {
-            perror("open");
-            return -1;
-        }
-        break;
-    case 'm':
-        mmonce = optarg;
-        is_mmonce_file = false;
-        break;
-    case 'n':
-        mmonce = optarg;
-        is_mmonce_file = true;
-        break;
-    }
-
-    if (mmonce == NULL) {
-        fprintf(stderr, "Error: no mmonce provided\n");
-        return 1;
-    }
-
-    domid = find_domain(argv[optind]);
-
-    rc = libxl_domain_attestation(ctx, domid, dst_file, is_mmonce_file, mmonce);
-
-    return rc;
-}
-
 /*
  * Local variables:
  * mode: C
